chore: move recurringd test out of devimint
